Special handling of check_recovery_conf for v12+
When starting as a replica it may take some time before Postgres starts
accepting new connections, but meanwhile it could happen that the leader
transitioned to a different member and the `primary_conninfo` must be
updated.

On pre v12 Patroni regularly checks `recovery.conf` in order to check
that recovery parameters match the expectation. Starting from v12
recovery parameters were converted to GUC's and Patroni gets current
values from the `pg_settings` view. The last one creates a problem
when it takes more than a minute for Postgres to start accepting new
connections.

Since Patroni attempts to execute at least pg_is_in_recovery() every
HA loop, and it is raising at exception, the `check_recovery_conf()`
effectively wasn't reachable until recovery is finished, but it
changed when #2082 was introduced.

As a result of #2082 we got the following behavior:
1. Up to v12 (not including) everything was working as expected
2. v12 and v13 - Patroni restarting Postgres after 1m of recovery
3. v14+ - the `check_recovery_conf()` is not executed because the
  `replay_paused()` method raising an exception.

In order to properly handle changes of recovery parameters or leader
transitioned to a different node on v12+, we will rely on the cached
values of recovery parameters until Postgres becomes ready to execute
queries.