Let patroni --validate-config exit 1 when config is invalid
Previously, Schema.__call__() did not return anything so the 'if error:'
clause in Config.__init__() never applied. Now it returns a list of
error messages, which is joined when raising the ConfigParseError. As a
result 'patroni --validate-config' now exits 1 when the configuration is
invalid.

In daemon.abstract_main(), we need to detect if --validate-config is on
in order not to display help message.