Implement failsafe_mode
If enabled it will allow Patroni to cope with DCS outages.
In case of DCS outage the leader tries to call all remaining members in
the cluster via API and if all of them respond with success the leader
will not be demoted.

The failsafe_mode could be enabled by running
```
patronictl edit-config -s failsafe_mode=true
```

or by calling the /config REST API endpoint.

Low-level implementation details.
---------------------------------

Since Patroni doesn't know anything about topology and how members are
distributed across different data centers or networks we make the worst
possible assumption, that all members are located in the same network,
while DCS is in fact could be distributed across different networks.
Whatever observed as DCS outage by the leader in fact could be a
partitioned network, but since the leader could see remaining members
demote will not happen. At the same time, the other member could be
started in a different network not affected by partitioning and we don't
want this member to promote because it will leader to a split-brain
situation. In order to protect from it, Patroni introduce a new key in
DCS, named /failsafe, where all cluster members are listed. The key is
maintained by the current leader of the Patroni cluster and only members
that are listed in it are allowed to participate in the leader race.