Don't drop replication slots in pause
If replication slots are enabled Patroni automatically creates them for
any cluster member that is supposed to stream from a given node and for
any permanent slot defined in the global configuration. If the member
disappears from the DCS Patroni automatically removes the replication
slot for it. The same behavior was in the maintenance mode (pause).

This commit disables removal of any replication slots that don't
match Patroni's expectations in pause.

Close https://github.com/zalando/patroni/issues/2314