Call pg_replication_slot_advance() from a thread
on busy clusters with many logical replication slots the
pg_replication_slot_advance () call affects the main HA loop and
could result in the member key expiration. The only way to solve it is
a dedicated thread response for moving slots forward.
The thread is started only when there are logical slots to be advanced.

Will help to solve #2388
Close #2239