Enforce search_path=pg_catalog for non-replication connections
There is a known vector of attact by creating functions and/or operators
in a public scheme with the same name and signature as corresponding
objects in pg_catalog. Since Patroni is heavily relying on superuser
connections we want to mitigate it by enforcing search_path=pg_catalog
for all connections created by Patroni (except replication connections).
It is achieved by introducing a new function, that wraps
psycopg.connect() function and appends '-c search_path=pg_catalog' to
`options` parameter.
In addition to that, we set connection.autocommit to True before
returning it.