Enable SSL in REST API and Postgres if possible when running behave
If openssl binary is available use it to generate a self-signed
certificate. Use it to protect Patroni REST API
(verify_client=required).

In case if Postgres is compiled with SSL support enable it in the
configuration and configure pg_hba.conf to check client certificates in
addition to passwords. Also configure superuser/replication/rewind users
to use client certificates and verify server certificate (verify-ca)