Suppress recurring errors when dropping unknown but active replication slots
When a replication slot is not registered with Patroni but is active, Patroni
would log an error during each HA cycle in certain conditions (after a restart
or role change). To avoid this, check if the replication slot was still active
if we fail to drop it and if so, only log a message at debug level. Otherwise,
log the slot we dropped for informational purposes.

Close #2499