Do not rely on 'role' value when checking other nodes via REST API
when doing the leader race we need to check that the former primary
isn't alive anymore. For that we relied on non-inclusive terms.
In order to simplify future work on getting rid from all non-inclusive
words we changed the check to rely on a difference in format of wal/xlog
field. There is only "location" for the primary and "replayed_location"
+ "received_location" for standbys.

In addition to that we start supporting "wal" field as well as
deprecated "xlog".