Implement Citus compatibility in DCS
Citus cluster (coordinator and workers) will be stored in DCS as a fleet
of Patroni clusters logically grouped together:
```
/service/batman/
/service/batman/0/
/service/batman/0/initialize
/service/batman/0/leader
/service/batman/0/members/
/service/batman/0/members/m1
/service/batman/0/members/m2
/service/batman/
/service/batman/1/
/service/batman/1/initialize
/service/batman/1/leader
/service/batman/1/members/
/service/batman/1/members/m1
/service/batman/1/members/m2
...
```

Where 0 is a Citus group for coordinator and 1, 2, etc are worker
groups.

Such hierarchy allows reading the entire Citus cluster with a single
call to DCS (except Zookeeper).

The get_cluster() method will be reading the entire Citus cluster on the
coordinator because it needs to discover workers. For the worker cluster
it will be reading the subtree of its own group.

Besides that we introduce a new method  get_citus_coordinator(). It will
be used only by worker clusters.

Since there is no hierarchical structures on K8s we will use the citus
group suffix on all objects that Patroni creates.
E.g.
```
batman-0-leader  # the leader config map for the coordinator
batman-0-config  # the config map holding initialize, config, and history "keys"
...
batman-1-leader  # the leader config map for worker group 1
batman-1-config
...
```