Implement Citus integration
Citus integration is enabled from patroni.yaml:
```yaml
citus:
  database: citus
  group: 0  # 0 is for coordinator, 1, 2, etc are for workers
```

If enabled, Patroni will create the database, citus extension in it, and
INSERTs INTO `pg_dist_authinfo` information required for Citus nodes to
communicate between each other, i.e. 'password', 'sslcert', 'sslkey' for
superuser if they are defined in the Patroni configuration file.

When the new Citus coordinator/worker is bootstrapped, Patroni adds
synchronous_mode=on to the `bootstrap.dcs` section.

Besides that, Patroni takes over management of some Postgres GUCs:
- shared_preload_libraries - Patroni ensures that the "citus" is added
  to the first place
- max_prepared_transactions - if not set or set to 0, Patroni changes
  the value to max_connections*2
- wal_level - automatically set to logical. It is used by Citus to
  move/split shards. Under the hood Citus is creating/removing
  replication slots and they are automatically added by Patroni to the
  `ignore_slots` configuration to avoid accidental removal.

The coordinator primary actively discovers worker primary nodes and
registers/updates them in the `pg_dist_node` table using
citus_add_node() and citus_update_node() functions.

Patroni running on the coordinator provides the new REST API endpoint:
`POST /citus`. It is used by workers to facilitate controlled
switchovers and restarts of worker primaries.
When the worker primary needs to shut down Postgres because of restart
or switchover, it calls the `POST /citus` endpoint on the coordinator
and the Patroni on the coordinator starts a transaction and calls
citus_update_node(nodeid, 'host-demoted', port) in order to pause client
connections that work with the given worker. Once the new leader is
elected or postgres started back, they perform another call to the `POST
/citus` endpoint, that does another citus_update_node() call with actual
hostname and port and commits a transaction.
After transaction is committed, coordinator reestablishes connections to
the worker node and client connections are unblocked. If clients don't
run long transaction the operation finishes without client visible
errors, but only a short latency spike.

All operations on the pg_dist_node are serialized by Patroni on the
coordinator. It allows to have more control and ROLLBACK transaction in
progress if its lifetime exceeding a certain threshold and there are
other worker nodes should be updated.