Run only one query per HA loop
If the cluster is stable (no nodes are joining/leaving/lagging) we want
to run at most one monitor query per every HA loop. So far it worker
perfectly except when synchronous_mode is enabled, where we run two
additional queries:
1. SHOW synchronous_mode
2. SELECT ... FROM pg_stat_replication

In order to solve it, we will include these "queries" to the common
monitoring query is synchronous_mode is enabled.

In addition to that we make sure that synchronous_standby_names is reset
on replicas that used to by a primary.

P.S.: in the monitoring query we also extract the current value of
synchronous_standby_names, because it will be useful for the quorum
commit feature.