Enhanced sync connections check
When `synchronous_standby_names` GUC is changed PostgreSQL nearly
immediately starts reporting corresponding walsenders as synchronous,
while in fact they maybe didn't reach this state yet. To mitigate this
problem we memorize current flush lsn on the primary right after change
of `synchronous_standby_names` got visible and use it as an additional
check for walsenders. The walsender will be counted as truly "sync" only
when write/flush/replay_lsn on it reached memorized LSN and the
`application_name` is known to be a part of `synchronous_standby_names`.