Use config file as a fallback when all current etcd nodes failed
If communication with etcd nodes failed it is logical to start from
scratch, from nodes that are listed in the config. But, it could happen
that config is in fact outdated and all nodes in the real cluster were
replaced.
Previously we used to track whether config file was changed, which
turned out not to work in all possible cases.
The new strategy is a bit more different - if communication with all
nodes failed we will continue keeping the last know topology and at the
same time will try to figure out the new one by merging two lists
together, the cached list and the list from the config file.