Don't write to PGDATA if major version is not known
It could happen that Patroni is started up before PGDATA was mounted.
In this case Patroni can't determine major Postgres version from
PG_VERSION file. Later, when PGDATA is mounted, Patroni was trying to
create the recovery.conf even if the actual Postgres major version is
newver than 12.
To mitigate the problem we double check that the
Postgresql._major_version is set and give up if it is not before writing
recovery configuration or starting postgres up.

Close https://github.com/zalando/patroni/issues/2434