Make sure Cluster.sync is never empty
It was possible to have it empty if the all cluster keys are missing in
DCS. In this case the Cluster object was manually created with all
values set to None or [] (including sync).
It already resulted in #2217, which is in fact wasn't a correct fix.

In order to solve it and reduce code duplication we introduce
Cluster.empty() and SyncState.empty() methods, which will create
corresponding empty objects and start using Cluster.empty() from all
places where empty Cluster object was manually created.