Remove `rint` custom function from `patroni/utils.py`
The `rint` function was created back then because of the difference
in `round` behavior between Python 2 and Python 3:

* Python 2: if two multiples are equally close, rounding is done away
from 0;
* Python 3: if two multiples are equally close, rounding is done toward
the even choice

Besides that Python 2 would return a `float` even if `ndigits` parameter
were `0`, while Python 3 would return an `int` in that case.

As Patroni no longer supports Python 2 we are removing `rint` from
the source code, and just using `round` instead.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>