Cover `etcd3` in `parse_dcs` function
Previous to this commit the `parse_dcs` function would fail with a
`PatroniCtlException` if the user ever passed an `etcd3` URL through
`--dcs-url` command-line option.

As a consequence, the only way of using `etcd3` in `patronictl` was
by using a configuration file passed through `-c` command-line option.

This commit fixes that issue and allows `etcd3` to be used in `--dcs-url`.

References: PAT-91 #2638

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>