Add postgres `initdb` method and modify argument handling
Adds an `unquote` function to `utils` to make sure literal quotes are
not sent to the initdb as configuration options. This is a problem for
any postgres setting that can take a value with spaces, such as a command
to run, e.g. --key-wrap-command in TDE enabled postgres variants.

References: PAT-86 #2632