Add some variable reuse to tox.ini
Note: in regard to defaults for `setenv` in `[testenv]`. If a specific env section
does not set `setenv` it will use the defaults. If any specific `[testenv<x>]`
section sets any key/value it overrides the default in `[testenv]`.

Therefore, the env var for OPEN_CMD is set in the defaults for `[testenv]setenv` whilst
env vars for other sections is set using a common section that is then referenced in
the env specific `setenv`. This is true also for other lists.

Unfortunately, it doesn't appear possible to reuse these variables in the section names
themselves, so duplication here is necessary.