Attempt to address a `codacy` issue
This commit attempts to fix this issue:

```
Number of parameters was 3 in 'BaseServer.handle_error' and is now 2 in overriding 'RestApiServer.handle_error' method
```

The method `handle_error` was overloaded, and the definition in the
superclass is an instance method, while the definition in Patroni's
`handle_error` was using `staticmethod`.

Had to adjust unit tests accordingly, so they now use an instance
of `MockRestApiServer` instead of a static call to `handle_error`.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>