Complement Postgres settings validation with `postgres --describe-config`
We now get all available GUCs from a given Postgres version by calling
`postgres --describe-config` command.

That list of GUCs is used for 2 purposes:

* Do not allow Postgres GUCs to be written to `postgresql.conf` or
`recovery.conf` if they do not exist in that Postgres version;
* Do not discard a Postgres GUC if it does not have a validator, but
it still exists in the Postgres version. The goal here is to avoid that
an eventual oversight in the Patroni YAML specs would cause a valid GUC
to be discarded. It is preferrable to not validate the GUC than to discard
it.

Besides that, this commit also changes the `transform_postgresql_parameter_value`
function so it attempts to find a validator for a GUC that contains a
`.` in its name. When a GUC contains a `.` it is very likely to be
a GUC from a Postgres extension. However, it may be the case that
a custom Postgres build contains a namespaced GUC. With that in mind
we check if the GUC contains a validator before just returning it as
it is (with no validation).

References: PAT-102.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>