Apply suggestions from code review
`_get_module` function was used for 2 purposes -- get references to:

1. Child classes of `_Transformable`
2. `parameters` and `recovery_parameters` variables

With this commit we drop the `_get_module` function.

References to child class of `_Transformable` are now gotten with a
dynamic dict which is built through the introduced method
`getsubclasses`.

References to `parameters` and `recovery_parameters` are now gotten
with a static dictionary.

Co-authored-by: Alexander Kukushkin <cyberdemn@gmail.com>