Add an exception for PG <= 11 when validating against `--describe-config` output.
PG <= 11 recovery parameters do not show up in the `--describe-config`
output.

Consenquently we cannot keep an strict check of "name in available_gucs".
From this commit onward we will allow that a PG <= 11 setting is not
present in `availble_gucs` (`--describe-config`) as long as Patroni
has defined a validator for it.

More details comments can be found in the code.

References: PAT-102.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>