Add validator for `postgresql.bin_name.*` and change `postgresql.bin_dir` validator.
This commit changes the Patroni config validator to allow the following
settings to be configured/validated:

* `postgresql.bin_name.pg_ctl`
* `postgresql.bin_name.initdb`
* `postgresql.bin_name.pg_controldata`
* `postgresql.bin_name.pg_basebackup`
* `postgresql.bin_name.postgres`
* `postgresql.bin_name.pg_isready`

The idea of these settings is to allow custom Postgres binary names
for the ones used by Patroni. That is usefull for example if someone
is running a custom build of Postgres which is delivered by a vendor.

Through this commit we create the function `get_bin_name`. That function
is used to get the configured value for a given `postgresql.bin_name`
entry, and if not configured then fall back to default binary name in a
Community Postgres distribution.

It also implements a new class `BinDirectory` which inherites from `Directory`.
The idea of this class is to validate the configured Postgres binary directory.
We needed to implement a new class so we could use the configured values for
`postgresql.bin_name`, if any, instead of just using the default binary names.

Besides that, this commit also fixes a few issues:

* `flake8` was complaining about a few formatting/docstrings things
* `validate_binary_name` fuction was missing a couple things:
    * to declare the expected value type
    * to raise `ConfigParserError` when issues were detected

References: PAT-17.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>