Use `postgresql.bin_name.*` if configured
This commit starts making use of these settings, if available:

* `postgresql.bin_name.pg_ctl`
* `postgresql.bin_name.initdb`
* `postgresql.bin_name.pg_controldata`
* `postgresql.bin_name.pg_basebackup`
* `postgresql.bin_name.postgres`
* `postgresql.bin_name.pg_isready`

The idea of these settings is to allow custom Postgres binary names
for the ones used by Patroni. That is usefull for example if someone
is running a custom build of Postgres which is delivered by a vendor.

We basically changed `pgcommand` method so it searches for available
`postgresql.bin_name`, and if not available fall back to default binary
name.

We also added `postgresql.bin_name.pg_rewind` which was missing in the
previous commit.

References: PAT-17.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>