Make sure the leader key isn't unconditionally overwritten
it may happen that the leader heart-beat loop didn't run properly and
the session has expired. In this case the leader may create a new
session and a new leader key. But, there are chances that the other node
already created a leader key and we don't want to overwrite it.