Refactor `_load_postgres_guc_validators`
We renamed the function `_load_postgres_guc_validators` to
`_get_postgres_guc_validators` and changed its behavior.

`_load_postgres_guc_validators` used to:

* Parse validators spec with `ValidatorFactory` and append them to
`parameters[parameter]` (or `recovery_parameters[parameter]`)
* Yield any exception that was faced, so the calling method would
log them

We had 2 problems with that approach:

* If a GUC spec is present both at `0_postgres.yml` and at a let's say
`1_vendor.yml`, it would create a resulting object with all validators
from both files
* Keep returning exceptions among functions so the caller could log a
more informative message, pointing the file which presented problems

Through this commit we change the behavior so:

* If a GUC is present at 2 or more YAML files, the last one takes precedence.
That seems a more appropriate behavior if for whatever reason a vendor changes
the behavior of a GUC
* The own function that observes the exception takes care of logging
a warning message. As we don't know the file at this point, we also
created a debug logging at `_load_postgres_gucs_validators` caller
function, so we can track that

References: PAT-102.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>