Fix a problem with etcd3.update_leader()
It didn't took into account the fact that we can get a new lease after
changing a TTL. In this case we have to update the exiting leader key
with the new lease.
To solve it we introduce the on transaction 'failure' callback.
The whole workflow looks like (schematically):
txn(
    compare=(old_value == self._name)),
    success=put(self.leader_path, self._name, self._lease),
    failure=txn(
        compare=(create_revision == '0'),
        success=put(self.leader_path, self._name, self._lease)
    )
)

The problem was introduced in d98d6d0b02c9cc67464a7b1b31b1c5570c26e12d