Remove `pg_rewind` from required binaries in validator.
Using `pg_rewind` through Patroni is optional. Also, Patroni supports
Postgres 9.3 and Postgres 9.4, which does not ship `pg_rewind` by
default and requires the user to compile it.

With that in mind we are removing `pg_rewind` from the strict list
of required binaries in `postgresql.bin_dir`.

References: PAT-17.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>