Adjust `CaseInsensitiveSet.__repr__` doctest
We had previously set the `SKIP` flag for that test based on the
assumption that `dict` keys querying order was not guaranteed based
on insertion order. However, that is not true anymore since Python
3.7.

The only method that should keep with `SKIP` is `__str__` as it uses
a `set` internally, which won't deterministically guarantee the values
order.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>