REST API would not reload SSL certificate upon receiving an SIGHUP
Patroni used to have basically 3 ways of loading a new certificate file
in the REST API server:

* Restarting the Patroni agent
* Changing the value of `restapi.certfile` and issuing an SIGHUP
* Changing the actual certificate pointed by `restapi.certfile` and issuing
an SIGHUP

The 3rd method relies on checking differences between serial numbers of
the configured certificate before and after an SIGHUP.

Through [#2648](https://github.com/zalando/patroni/pull/2648) we applied
a change to method `get_certificate_serial_number` that introduced a bug:
REST API's certificate will not be reloaded upon an SIGHUP if the value
of `restapi.certfile` remains unchanged, but with a different content in
the actual certificate file.

We are currently attempting the use the output of `load_verify_locations`
to get the serial number of the certificate. However, that method only
loads the information, and we need to fetch the serial number from
entries returned by `get_ca_certs` instead.

This commit applies the above fix. Unit tests have been changed accordingly.

References: PAT-130.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>