Start postgres not in recovery in some cases
If we know for sure that a few moments ago postgres was still running as
a primary and we still have the leader lock and can successfully update
it, in this case we can safely start postgres back not in recovery.
That will allow to avoid bumping timeline without a reason and hopefully
improve reliability because it will address issues similar to #2720.

In addition to that remove `if self.state_handler.is_starting()` check
from the `recover()` method. This branch could never be reached because
the `starting` state is handled earlier in the `_run_cycle()`.
Besides that remove redundant `self._crash_recovery_executed`.

P.S. now we do not cover cases when Patroni was killed along with
Postgres.
Lets consider that we just started Patroni, there is no leader, and
`pg_controldata` reports `Database cluster state` as `shut down`.
It feels logical to use `Latest checkpoint location` and
`Latest checkpoint's TimeLineID` to do a usual leader race and start
directly as a primary, but it could be totally wrong. The thing is that
we run `postgres --single` if standby wasn't shut down cleanly before
executing `pg_rewind`. As a result `Database cluster state` transition
from `in archive recovery` to `shut down`, but if such a node becomes a
leader the timeline must be increased.