Make sure the version_prefix for etcd3 is set to /v3beta
Setting ot before calling the parent constructor didn't really work
because it is being overwritten in the `etcd.Client.__init__()`.

The only viable way of doing it passing the custom value to the parent
class. As a side effect it becomes possible to change `version_prefix`
from the `etcd` configuration section (but not for etcd3!), but we will
not document it.

Close https://github.com/zalando/patroni/issues/2142