Allow integer gucs without units in validation
Previously, integer gucs, for example `max_connections` would not pass
the validation, as these settings have no unit, if and only if they were
specified as a string.

This causes problems if the `max_connections` is configured in
`patroni.yaml` as a string, for example, the following configuration
would not result in the right `max_connections` settings, as
`max_connections` is configured as a string:

    bootstrap:
      dcs:
        postgresql:
          parameters:
            log_checkpoints: "on"
            log_connections: "off"
            max_connections: "57"

Allowing a user to specify *all* parameters as a string was accepted
before in Patroni and also seems very useful, as many of us will be
using Ansible/Helm/Golang to build a Patroni configuration, in which
creating a `map[string]string` is easier than having to deal with data
types.

Attempts to address issue #2735