Do not fail validation for a value that is fine
We fail validation for parameters that have the _false_validator,
however, if a user specifies these values themselves and they match the
required value, we shouldn't fail the validation - the value is actually
ok.

An example was for a configuration system that explicitly set
`hot_standby=on` in the PostgreSQL configuration. This *failed*
validation previously.

To ensure users can understand why configuration values may differ from
what they specified, we now also log this when writing the
configuration.