Take client certificates only from ctl section
or from respective environment variables.

Motivation: sometimes there are server-only certificates that can't
be used as client certificates. As a result neither Patroni not
patronictl work correctly even if ``--insecure`` option is used.

Reflect these changes in the documentation and also add a warning that
if ``restapi.verify_client`` set to ``required`` then client
certificates must be provided in the ``ctl`` section.