Create a new contributing level in docs toctree
The existing "Contributing guidelines" are moved into their own
file and a new "CONTRIBUTING" page has been created with a
`toctree` which adds this and places API docs beneath.

As the `maxdepth` of the index `toctree` is 2 the subsections on
the Contributing guidelines page are no longer visible, they are
shown on the new CONTRIBUTING `toctree` instead.