Set permissions for files and directories created in PGDATA
Postgres supports two types of permissions:
1. owner only
2. group readable

By default the first one is used because it provides better security.
But, sometimes people want to run a backup tool with the user that is
different from postgres. In this case the second option becomes very
useful. Unfortunately it didn't work correctly because Patroni was
creating files with owner access only permissions.

This PR changes the behaviour and permissions on files and directories
that are created by Patroni will be calculated based on permissions of
PGDATA. I.e., they will get group readable access when it is necessary.

Close #1899
Close #1901